相关接口下线： 1、根据用户ID,获取用户来自的第三方平台（10001012） 2、兑换礼物（10005054） 3、贴标签(20010201) 4、用户删除被贴标签(20010202) 5、获取指定类别的座驾信息接口(10005056) 6、获取用户送阳光中奖类别和详细明细信息接口（10009012） 7、获取新注册任务列表(10005020) 8、完成任务列表中的任务(10005021) 9、7日签到补签(50005002) 10、7日签到抽奖(50005003) 11、7日签到抽奖资格校验(50005004) 12、领取任务奖励(10005022) 13、获取用户中奖列表（10005024） 14、获取通告列表(10002007) 15、获取公告详细(10002012) 16、获取活动详细(10002013) 17、获取用户一定时间内收到礼物个数(10002041) 18、获取一定时间内礼物送出总数、打折总数和用户总数(10002042) 19、移动端体育精选频道 (20010310) 20、用户分享推荐 [fucTag=20040008] 21、获取指定礼物指定时间内按一定比例获取到的消耗值（10007009） 22、购买表情包(20000007) 23、获取房间礼物特效接口（50001014） 24、获取用户消费记录列表（10005006） 25、获取指定礼物指定时间内按一定比例获取到的消耗值（10007007） 26、授权管理（10003009） 27、收回管理权限（10003010）
去Oracle相关接口调整：
28、获取用户送出的礼物列表（10005007）
29、获取用户收到的礼物列表（10006001）
30、获取用户粉丝列表(10003004) 去Oracle
31、获取用户信息(For Node)(10005044)
32、获取用户信息(For Node)(10005044) welcomeMsg，noticeContent，welcomeMsgHref相关属性调整
33、获取房间信息(For Node)(10001052) welcomeMsg，noticeContent，welcomeMsgHref相关属性调整
34、获取本周礼物消耗排行榜(10002011)
35、获取用户本周上榜礼物列表(10005025)
36、获取上周周星礼物第一名排行榜(10001044)
37、获取本周周星礼物(10001045)
38、获取新版主播周星争夺战(10002074)
39、获取周星主播和用户前三名排行榜(10002073)
40、上传用户设备信息接口(10007003)
41、获取主播个性礼物（50001005）
42、添加主播个性礼物（50001006）
43、删除主播个性礼物（50001007）
44、删除已下架的用户个性礼物（50001008）
45、发送短信（40000020）
46、获取申请家族信息（50008028）
47、手机注册（40000011）

+++ src/main/java/com/melot/kkcx/transform/SystemMessageTF.java
@@ -30,7 +30,7 @@ public class SystemMessageTF {
             return null;
         }
         KkSystemNotice kkSystemNotice = new KkSystemNotice();
-        kkSystemNotice.setId(systemMessage.getId());
+        kkSystemNotice.setId((long)systemMessage.getId());
         kkSystemNotice.setDescribe(systemMessage.getDescribe());
         kkSystemNotice.setMsgtime(systemMessage.getTime());
         kkSystemNotice.setTitle(systemMessage.getTitle());
